[BE] FIX : 전체랭킹 수정

+++ BE/gumid101/src/main/java/com/ssafy/gumid101/res/RankingDto.java
@@ -7,5 +7,5 @@ public class RankingDto {
 
 	private String userName;
 	private Long userSeq;
-	private Long lankingValue;
+	private Integer lankingValue;
 }
